+++ milvus-20231012/alibabacloud_milvus20231012/client.py
@@ -21,17 +21,6 @@ class Client(OpenApiClient):
     ):
         super().__init__(config)
         self._endpoint_rule = 'regional'
-        self._endpoint_map = {
-            'eu-central-1': 'milvus.eu-central-1.aliyuncs.com',
-            'cn-zhangjiakou': 'milvus.cn-zhangjiakou.aliyuncs.com',
-            'cn-wulanchabu': 'milvus.cn-wulanchabu.aliyuncs.com',
-            'cn-shenzhen': 'milvus.cn-shenzhen.aliyuncs.com',
-            'cn-shanghai': 'milvus.cn-shanghai.aliyuncs.com',
-            'cn-hongkong': 'milvus.cn-hongkong.aliyuncs.com',
-            'cn-hangzhou': 'milvus.cn-hangzhou.aliyuncs.com',
-            'cn-beijing': 'milvus.cn-beijing.aliyuncs.com',
-            'ap-southeast-1': 'milvus.ap-southeast-1.aliyuncs.com'
-        }
         self.check_config(config)
         self._endpoint = self.get_endpoint('milvus', self._region_id, self._endpoint_rule, self._network, self._suffix, self._endpoint_map, self._endpoint)
 
@@ -334,6 +323,8 @@ class Client(OpenApiClient):
             body['loadReplicas'] = request.load_replicas
         if not DaraCore.is_null(request.multi_zone_mode):
             body['multiZoneMode'] = request.multi_zone_mode
+        if not DaraCore.is_null(request.node_type):
+            body['nodeType'] = request.node_type
         if not DaraCore.is_null(request.payment_duration):
             body['paymentDuration'] = request.payment_duration
         if not DaraCore.is_null(request.payment_duration_unit):
@@ -418,6 +409,8 @@ class Client(OpenApiClient):
             body['loadReplicas'] = request.load_replicas
         if not DaraCore.is_null(request.multi_zone_mode):
             body['multiZoneMode'] = request.multi_zone_mode
+        if not DaraCore.is_null(request.node_type):
+            body['nodeType'] = request.node_type
         if not DaraCore.is_null(request.payment_duration):
             body['paymentDuration'] = request.payment_duration
         if not DaraCore.is_null(request.payment_duration_unit):